requirements.txt
Django==2.1.4
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
six==1.12.0
